Add files via upload
The Notes Taking App is a simple web application that allows users to create, view, and store notes using JavaScript, HTML, and CSS. Users can type their notes into a text area and add them to a list. The app leverages the browser's local storage to ensure that notes persist even after the page is refreshed. Its user-friendly interface is designed for easy interaction, making it a great tool for jotting down thoughts, ideas, or tasks on the go. Additionally, the app can be enhanced with features like note editing, deletion, and search functionality to improve user experience.